Add `release_label` configuration
Some repositories may use an automated tool or bot to merge their pull
requests. In this case the bot could be triggered by labeling the pull
request instead of merging directly to the base branch. This allows
a `release_label` to be configured which will result in `git release`
adding the label rather than merging to the base branch.