Update local main branch during `git update`
This ensures that the local main branch is updated during `git update`.
Without this it's necessary for the end user to run `git update`
separately on the main branch, and it can be confusing if they e.g. want
to diff their newly-updated feature branch against the main branch and
don't realize it's stale.