Don't confirm branch status when `release_label` configured
At least once every couple weeks I think I released some code via `git
release` and come back to my console hours later (or the next day) to
find it sitting on the `Branch status is currently: pending.  Proceed
with release? (y/n)` prompt.

This prompt is really helpful when there isn't too much churn on the
main branch, but with high churn it's likely to show up a lot more
often. Moreover, it's less important when using a merge tool via
`release_label` because `gitx` isn't merging the code directly and the
merge tool will presumably be confirming the build status before merging
the code.

This seemed like a pragmatic (if not a little magic) change. I could see
an alternative to this being the ability to suppress the prompt via
a separate config variable instead of automatically when `release_label`
is used.