Mitigate post-release race condition with pull request
Since we've switched over to using `release_branch`, sometimes the merge
tool will merge and close the pull request between when `git release`
adds the label and when it runs `git integrate`. This causes `gitx` to
not be able to find the PR (since it's no longer `open`) and prompts the
user to create one.

This works around that issue by adding a `--no-pull-request` option to
`git integrate` to skip the pull request reference, and specifying that
option in the `after_release` default config.