Add configuration to skip update during release
When a repository uses a merge tool, the merge tool typically already
performs and update from the base branch as part of evaluating the
changeset. In this case updating from the base branch locally as part of
the release process can actually add latency to the release process,
since CI will typically have to run again after the merge before the
changeset can enter the merge queue.

This change adds a new `update_from_base_on_release` configuration that
can be set to `false` to bypass the update.