Pull without rebase
As described in https://github.com/wireframe/gitx/issues/52, when the
developer has `pull.rebase = true` in their git config, they will get an
error when updating due to git attempting to rebase local changes on top
of a different remote HEAD. The error can unfortunately be fairly opaque
and confusing (e.g. `Updates were rejected because the tip of your
current branch is behind`). This fixes the issue by doing a plain merge
instead of a rebase.